feat: allow additional ddns-* options
`ddns-rev-domainname` is added at the top level of the configuration and
in the subnet section too.
`ddns-domainname` is added to the subnet section as well because it's
possible to do so in configuration (ie. supose that you want hosts in
 a given subnet to have a different domain, for instance iot.foo.com).

`ddns-hostname` is added for host section because actually quite handy
to fix broken names that some client are generating.